requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cloudinary==1.44.2
cryptography==47.0.0
Django==5.0.14
django-allauth==65.16.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.13
packaging==26.2
pillow==10.4.0
psycopg2-binary==2.9.12
pycparser==3.0
PyJWT==2.12.1
razorpay==1.4.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
urllib3==2.6.3
whitenoise==6.12.0
